+++ CHANGELOG.md
@@ -19,6 +19,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - CaseS HPO term searches for multiple terms works independent of order
 - CaseS search regexp should not allow backslash
 - CaseS cohort tags can contain whitespace and still match
+- Remove diagnoses from cases even if OMIM term is not found in the database
 - Parsing of HPO terms associated to OMIM diagnoses
 ### Changed
 - Column width adjustment on caseS page